@@ -14,15 +14,14 @@
  * License along with this library; if not, write to the Free Software
  * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA 02111-1307 USA
  *
- * $Id$
  */
 
 package org.kopi.ebics.certificate;
 
+import java.io.File;
 import java.io.FileInputStream;
 import java.io.IOException;
 import java.io.InputStream;
-import java.io.InputStreamReader;
 import java.io.OutputStream;
 import java.math.BigInteger;
 import java.security.GeneralSecurityException;
@@ -43,7 +42,6 @@ import java.util.HashMap;
 import java.util.Map;
 import java.util.logging.Level;
 import java.util.logging.Logger;
-import org.bouncycastle.openssl.PEMReader;
 
 /**
  * Key store loader. This class loads a key store from
@@ -51,7 +49,6 @@ import org.bouncycastle.openssl.PEMReader;
  * for a given alias.
  * The PKCS12 key store type is recommended to be used
  *
- * @author hachani
  *
  */
 public class KeyStoreManager {
@@ -81,14 +78,11 @@ public class KeyStoreManager {
    * @throws GeneralSecurityException
    */
   public final PrivateKey getPrivateKey(String alias) throws GeneralSecurityException {
-    PrivateKey			key;
-
-    key = (PrivateKey) keyStore.getKey(alias, password);
-    if (key == null) {
-      throw new IllegalArgumentException("private key not found for alias " + alias);
-    }
-
-    return key;
+      PrivateKey key = (PrivateKey) keyStore.getKey(alias, password);
+      if (key == null) {
+          throw new IllegalArgumentException("private key not found for alias " + alias);
+      }
+      return key;
   }
 
   /**
@@ -98,7 +92,7 @@ public class KeyStoreManager {
    * @throws GeneralSecurityException
    * @throws IOException
    */
-  public void load(String path, char[] password)
+  public void load(File path, char[] password)
     throws GeneralSecurityException, IOException
   {
     keyStore = KeyStore.getInstance("PKCS12", "BC");
@@ -112,8 +106,8 @@ public class KeyStoreManager {
    * @throws GeneralSecurityException
    * @throws IOException
    */
-  private void load(String path) throws GeneralSecurityException, IOException {
-    if (path.equals("")) {
+  private void load(File path) throws GeneralSecurityException, IOException {
+    if (path == null) {
       this.keyStore.load(null, null);
     } else {
       this.keyStore.load(new FileInputStream(path), password);
@@ -127,20 +121,11 @@ public class KeyStoreManager {
    * @param provider the certificate provider
    * @return the certificate
    * @throws CertificateException
-   * @throws IOException
    */
   public X509Certificate read(InputStream input, Provider provider)
-    throws CertificateException, IOException
-  {
-    X509Certificate		certificate;
-
-    certificate = (X509Certificate) CertificateFactory.getInstance("X.509", provider).generateCertificate(input);
-
-    if (certificate == null) {
-      certificate = (X509Certificate)(new PEMReader(new InputStreamReader(input))).readObject();
-    }
-
-    return certificate;
+      throws CertificateException {
+      return (X509Certificate) CertificateFactory.getInstance("X.509",
+          provider).generateCertificate(input);
   }
 
   /**
@@ -148,15 +133,11 @@ public class KeyStoreManager {
    * @param input the given certificate
    * @return The RSA public key of the given certificate
    * @throws GeneralSecurityException
-   * @throws IOException
    */
-  public RSAPublicKey getPublicKey(InputStream input)
-    throws GeneralSecurityException, IOException
-  {
-    X509Certificate		cert;
+  public RSAPublicKey getPublicKey(InputStream input) throws GeneralSecurityException {
 
-    cert = read(input, keyStore.getProvider());
-    return (RSAPublicKey) cert.getPublicKey();
+      X509Certificate cert = read(input, keyStore.getProvider());
+      return (RSAPublicKey) cert.getPublicKey();
   }
 
   public RSAPublicKey getPublicKey(BigInteger publicExponent, BigInteger modulus)
@@ -168,7 +149,7 @@ public class KeyStoreManager {
             return null;
       }
   }
-  
+
   /**
    * Writes the given certificate into the key store.
    * @param alias the certificate alias
@@ -191,7 +172,7 @@ public class KeyStoreManager {
   {
     keyStore.store(output, password);
   }
-  
+
   /**
    * Returns the certificates contained in the key store.
    * @return the certificates contained in the key store.
@@ -207,29 +188,21 @@ public class KeyStoreManager {
    *         the key of the map is the certificate alias
    * @throws KeyStoreException
    */
-  public Map<String, X509Certificate> read(KeyStore keyStore)
-    throws KeyStoreException
-  {
-    Map<String, X509Certificate>	certificates;
-    Enumeration<String> 		enumeration;
-
-    certificates = new HashMap<String, X509Certificate>();
-    enumeration = keyStore.aliases();
-    while (enumeration.hasMoreElements()) {
-      String		alias;
-
-      alias = enumeration.nextElement();
-      certificates.put(alias, (X509Certificate)keyStore.getCertificate(alias));
-    }
-
-    return certificates;
+  public Map<String, X509Certificate> read(KeyStore keyStore) throws KeyStoreException {
+      Map<String, X509Certificate> certificates = new HashMap<>();
+      Enumeration<String> enumeration = keyStore.aliases();
+      while (enumeration.hasMoreElements()) {
+          String alias = enumeration.nextElement();
+          certificates.put(alias, (X509Certificate) keyStore.getCertificate(alias));
+      }
+      return certificates;
   }
 
   // --------------------------------------------------------------------
   // DATA MEMBERS
   // --------------------------------------------------------------------
 
-  private KeyStore			keyStore;
-  private char[]			password;
-  private Map<String, X509Certificate>	certs;
+    private KeyStore keyStore;
+    private char[] password;
+    private Map<String, X509Certificate> certs;
 }
