@@ -14,12 +14,11 @@
  * License along with this library; if not, write to the Free Software
  * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA 02111-1307 USA
  *
- * $Id$
  */
 
 package org.kopi.ebics.certificate;
 
-import java.io.UnsupportedEncodingException;
+import java.nio.charset.StandardCharsets;
 import java.security.GeneralSecurityException;
 import java.security.KeyPair;
 import java.security.KeyPairGenerator;
@@ -28,17 +27,16 @@ import java.security.NoSuchAlgorithmException;
 import java.security.SecureRandom;
 import java.security.interfaces.RSAPublicKey;
 
-import org.apache.commons.codec.binary.Base64;
 import org.apache.commons.codec.binary.Hex;
 import org.kopi.ebics.exception.EbicsException;
+import org.kopi.ebics.utils.Utils;
 
 /**
  * Some key utilities
  *
- * @author hachani
  *
  */
-public class KeyUtil {
+public final class KeyUtil {
 
     private KeyUtil() {
     }
@@ -50,34 +48,14 @@ public class KeyUtil {
    * @return KeyPair the key pair
    * @throws NoSuchAlgorithmException
    */
-  public static KeyPair makeKeyPair(int keyLen) throws NoSuchAlgorithmException{
+  public static KeyPair makeKeyPair(int keyLen) throws NoSuchAlgorithmException {
     KeyPairGenerator 		keyGen;
 
     keyGen = KeyPairGenerator.getInstance("RSA");
-    keyGen.initialize(keyLen, new SecureRandom());
+    keyGen.initialize(keyLen, Utils.secureRandom);
 
-    KeyPair keypair = keyGen.generateKeyPair();
+    return keyGen.generateKeyPair();
 
-    return keypair;
-
-  }
-
-  /**
-   * Generates a random password
-   *
-   * @return the password
-   */
-  public static String generatePassword() {
-    SecureRandom 		random;
-
-    try {
-      random = SecureRandom.getInstance("SHA1PRNG");
-      String pwd = Base64.encodeBase64String(random.generateSeed(5));
-
-      return pwd.substring(0, pwd.length() - 2);
-    } catch (NoSuchAlgorithmException e) {
-      throw new RuntimeException(e);
-    }
   }
 
   /**
@@ -111,14 +89,13 @@ public class KeyUtil {
     }
 
     try {
-      digest = MessageDigest.getInstance("SHA-256", "BC").digest(hash.getBytes("US-ASCII"));
+      digest = MessageDigest.getInstance("SHA-256", "BC").digest(hash.getBytes(
+          StandardCharsets.US_ASCII));
     } catch (GeneralSecurityException e) {
-      throw new EbicsException(e.getMessage());
-    } catch (UnsupportedEncodingException e) {
       throw new EbicsException(e.getMessage());
     }
 
-    return new String(Hex.encodeHex(digest, false)).getBytes();
+      return new String(Hex.encodeHex(digest, false)).getBytes();
   }
 
   /**
