@@ -14,7 +14,6 @@
  * License along with this library; if not, write to the Free Software
  * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA 02111-1307 USA
  *
- * $Id$
  */
 
 package org.kopi.ebics.exception;
@@ -22,20 +21,16 @@ package org.kopi.ebics.exception;
 /**
  * Download operation failed due to lack of data.
  *
- * @author hachani
  *
  */
 public class NoDownloadDataAvailableException extends EbicsException {
+    public NoDownloadDataAvailableException() {
+        super(ReturnCode.EBICS_NO_DOWNLOAD_DATA_AVAILABLE);
+    }
+    public NoDownloadDataAvailableException(String message) {
+        super(ReturnCode.EBICS_NO_DOWNLOAD_DATA_AVAILABLE, message);
+    }
 
-  @Override
-  public ReturnCode getReturnCode() {
-    return ReturnCode.EBICS_NO_DOWNLOAD_DATA_AVAILABLE;
-  }
-
-  @Override
-  public String getMessage() {
-    return getReturnCode().getText();
-  }
   // --------------------------------------------------------------------
   // DATA MEMBERS
   // --------------------------------------------------------------------
