@@ -14,18 +14,17 @@
  * License along with this library; if not, write to the Free Software
  * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA 02111-1307 USA
  *
- * $Id$
  */
 
 package org.kopi.ebics.interfaces;
 
+import java.io.File;
 import java.util.Locale;
 
 
 /**
  * EBICS client application configuration.
  *
- * @author hachani
  *
  */
 public interface Configuration {
@@ -34,25 +33,13 @@ public interface Configuration {
    * Returns the root directory of the client application.
    * @return the root directory of the client application.
    */
-  public String getRootDirectory();
-
-  /**
-   * Returns the default log directory of the application.
-   * @return the default log directory
-   */
-  public String getLogDirectory();
-
-  /**
-   * Returns the default log file name of the application
-   * @return the default log file name
-   */
-  public String getLogFileName();
+  File getRootDirectory();
 
   /**
    * Returns the EBICS configuration file.
    * @return the EBICS configuration file.
    */
-  public String getConfigurationFile();
+  File getConfigurationFile();
 
   /**
    * Returns the property value of a given key from
@@ -60,15 +47,15 @@ public interface Configuration {
    * @param key the given key
    * @return the property value
    */
-  public String getProperty(String key);
+  String getProperty(String key);
 
   /**
    * Returns the directory path of the key store that contains
    * bank and user certificates.
-   * @param the ebics user.
+   * @param user the ebics user.
    * @return the key store directory of a given user.
    */
-  public String getKeystoreDirectory(EbicsUser user);
+  File getKeystoreDirectory(EbicsUser user);
 
   /**
    * Returns the directory path that contains the traces
@@ -76,104 +63,97 @@ public interface Configuration {
    * @param user the ebics user
    * @return the transfer trace directory
    */
-  public String getTransferTraceDirectory(EbicsUser user);
+  File getTransferTraceDirectory(EbicsUser user);
 
   /**
    * Returns the object serialization directory.
-   * @param user the ebics user
    * @return the object serialization directory.
    */
-  public String getSerializationDirectory();
+  File getSerializationDirectory();
 
   /**
    * Returns the SSL trusted store directory.
    * @return the SSL trusted store directory.
    */
-  public String getSSLTrustedStoreDirectory();
+  File getSSLTrustedStoreDirectory();
 
   /**
    * Return the SSL key store directory
    * @return the SSL key store directory
    */
-  public String getSSLKeyStoreDirectory();
+  File getSSLKeyStoreDirectory();
 
   /**
    * Returns the SSL bank server certificates.
    * @return the SSL bank server certificates.
    */
-  public String getSSLBankCertificates();
+  File getSSLBankCertificates();
 
   /**
    * Returns the users directory.
    * @return the users directory.
    */
-  public String getUsersDirectory();
+  File getUsersDirectory();
 
   /**
    * Returns the Ebics client serialization manager.
    * @return the Ebics client serialization manager.
    */
-  public SerializationManager getSerializationManager();
+  SerializationManager getSerializationManager();
 
   /**
    * Returns the Ebics client trace manager.
    * @return the Ebics client trace manager.
    */
-  public TraceManager getTraceManager();
+  TraceManager getTraceManager();
 
   /**
    * Returns the letter manager.
    * @return the letter manager.
    */
-  public LetterManager getLetterManager();
+  LetterManager getLetterManager();
 
   /**
    * Returns the initializations letters directory.
    * @return the initializations letters directory.
    */
-  public String getLettersDirectory(EbicsUser user);
+  File getLettersDirectory(EbicsUser user);
 
   /**
    * Returns the users directory.
    * @return the users directory.
    */
-  public String getUserDirectory(EbicsUser user);
-
-  /**
-   * Returns the client application logger.
-   * @return the client application logger.
-   */
-  public EbicsLogger getLogger();
+  File getUserDirectory(EbicsUser user);
 
   /**
    * Configuration initialization.
    * Creates the necessary directories for the ebics configuration.
    */
-  public void init();
+  void init();
 
   /**
    * Returns the application locale.
    * @return the application locale.
    */
-  public Locale getLocale();
+  Locale getLocale();
 
   /**
    * Returns the client application signature version
    * @return the signature version
    */
-  public String getSignatureVersion();
+  String getSignatureVersion();
 
   /**
    * Returns the client application authentication version
    * @return the authentication version
    */
-  public String getAuthenticationVersion();
+  String getAuthenticationVersion();
 
   /**
    * Returns the client application encryption version
    * @return the encryption version
    */
-  public String getEncryptionVersion();
+  String getEncryptionVersion();
 
   /**
    * Tells if the client application should keep XML transfer
@@ -181,7 +161,7 @@ public interface Configuration {
    * @return True if the client application should not delete
    *         the XML transfer files
    */
-  public boolean isTraceEnabled();
+  boolean isTraceEnabled();
 
   /**
    * Returns if the files to be transferred should be
@@ -191,17 +171,17 @@ public interface Configuration {
    *
    * @return true if the file compression is enabled
    */
-  public boolean isCompressionEnabled();
+  boolean isCompressionEnabled();
 
   /**
    * Returns the default revision of sent XML.
    * @return the default revision of sent XML.
    */
-  public int getRevision();
+  int getRevision();
 
   /**
    * Returns the version of the EBICS protocol used by the client.
    * @return the version of the EBICS protocol.
    */
-  public String getVersion();
+  String getVersion();
 }
