@@ -14,7 +14,6 @@
  * License along with this library; if not, write to the Free Software
  * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA 02111-1307 USA
  *
- * $Id$
  */
 
 package org.kopi.ebics.interfaces;
@@ -26,7 +25,6 @@ import java.security.interfaces.RSAPublicKey;
 /**
  * Details about EBICS communication with a given bank.
  *
- * @author Hachani
  *
  */
 public interface EbicsBank extends Serializable {
@@ -35,55 +33,55 @@ public interface EbicsBank extends Serializable {
    * Returns the URL needed for communication to the bank.
    * @return the URL needed for communication to the bank.
    */
-  public URL getURL();
+  URL getURL();
 
   /**
    * 
    */
-  public boolean useCertificate();
+  boolean useCertificate();
   
   /**
    * 
    */
-  public void setUseCertificate(boolean useCertificate);
+  void setUseCertificate(boolean useCertificate);
   
   /**
    * Returns the encryption key digest you have obtained from the bank.
    * Ensure that nobody was able to modify the digest on its way from the bank to you.
    * @return the encryption key digest you have obtained from the bank.
    */
-  public byte[] getE002Digest();
+  byte[] getE002Digest();
 
   /**
    * Returns the authentication key digest you have obtained from the bank.
    * Ensure that nobody was able to modify the digest on its way from the bank to you.
    * @return the authentication key digest you have obtained from the bank.
    */
-  public byte[] getX002Digest();
+  byte[] getX002Digest();
 
   /**
    * Returns the banks encryption key.
    * @return the banks encryption key.
    */
-  public RSAPublicKey getE002Key();
+  RSAPublicKey getE002Key();
 
   /**
    * Returns the banks authentication key.
    * @return the banks authentication key.
    */
-  public RSAPublicKey getX002Key();
+  RSAPublicKey getX002Key();
 
   /**
    * Returns the bank's id.
    * @return the bank's id.
    */
-  public String getHostId();
+  String getHostId();
 
   /**
    * Returns the bank's name.
    * @return the bank's name.
    */
-  public String getName();
+  String getName();
 
   /**
    * Keys have been fetched from the bank.
@@ -93,12 +91,12 @@ public interface EbicsBank extends Serializable {
    * @param e002Key the banks encryption key.
    * @param x002Key the banks authentication key.
    */
-  public void setBankKeys(RSAPublicKey e002Key, RSAPublicKey x002Key);
+  void setBankKeys(RSAPublicKey e002Key, RSAPublicKey x002Key);
 
   /**
    * Sets the bank digests.
    * @param e002Digest encryption digest
    * @param x002Digest authentication digest
    */
-  public void setDigests(byte[] e002Digest, byte[] x002Digest);
+  void setDigests(byte[] e002Digest, byte[] x002Digest);
 }
