@@ -14,7 +14,6 @@
  * License along with this library; if not, write to the Free Software
  * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA 02111-1307 USA
  *
- * $Id$
  */
 
 package org.kopi.ebics.interfaces;
@@ -22,7 +21,6 @@ package org.kopi.ebics.interfaces;
 /**
  * Information about an EBICS customer.
  *
- * @author Hachani
  *
  */
 public interface EbicsPartner {
@@ -31,17 +29,17 @@ public interface EbicsPartner {
    * Returns the bank we are customer of.
    * @return the bank we are customer of
    */
-  public EbicsBank getBank();
+  EbicsBank getBank();
 
   /**
    * Returns the customers id at the bank.
    * @return the customers id at the bank.
    */
-  public String getPartnerId();
+  String getPartnerId();
 
   /**
    * Creates the next order number.
    * @return the next order number.
    */
-  public String nextOrderId();
+  String nextOrderId();
 }
