@@ -14,7 +14,6 @@
  * License along with this library; if not, write to the Free Software
  * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA 02111-1307 USA
  *
- * $Id$
  */
 
 package org.kopi.ebics.interfaces;
@@ -27,7 +26,6 @@ import org.kopi.ebics.exception.EbicsException;
 /**
  * An Ebics root element knows its name.
  *
- * @author hachani
  *
  */
 public interface EbicsRootElement extends EbicsElement {
@@ -36,26 +34,26 @@ public interface EbicsRootElement extends EbicsElement {
    * Converts the <code>EbicsElement</code> to a byte array
    * @return the equivalent byte array of this <code>EbicsElement</code>
    */
-  public byte[] toByteArray();
+  byte[] toByteArray();
 
   /**
    * Validates the request element according to the
    * EBICS XML schema specification
    * @throws EbicsException throws an EbicsException when validation fails
    */
-  public void validate() throws EbicsException;
+  void validate() throws EbicsException;
 
   /**
    * Adds a namespace declaration for the <code>EbicsRootElement</code>
    * @param prefix namespace prefix
    * @param uri namespace uri
    */
-  public void addNamespaceDecl(String prefix, String uri);
+  void addNamespaceDecl(String prefix, String uri);
 
   /**
    * Saves the <code>EbicsElement</code> into a given output stream.
    * @param out the output stream
    * @throws EbicsException the save operation fails
    */
-  public void save(OutputStream out) throws EbicsException;
+  void save(OutputStream out) throws EbicsException;
 }
