@@ -14,7 +14,6 @@
  * License along with this library; if not, write to the Free Software
  * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA 02111-1307 USA
  *
- * $Id$
  */
 
 package org.kopi.ebics.interfaces;
@@ -31,7 +30,6 @@ import org.kopi.ebics.exception.EbicsException;
 /**
  * Things an EBICS user must be able to perform.
  *
- * @author Hachani
  *
  */
 public interface EbicsUser {
@@ -40,112 +38,112 @@ public interface EbicsUser {
    * Returns the public part of the signature key.
    * @return the public part of the signature key.
    */
-  public RSAPublicKey getA005PublicKey();
+  RSAPublicKey getA005PublicKey();
 
   /**
    * Returns the public part of the encryption key.
    * @return the public part of the encryption key.
    */
-  public RSAPublicKey getE002PublicKey();
+  RSAPublicKey getE002PublicKey();
 
   /**
    * Return the public part of the transport authentication key.
    * @return the public part of the transport authentication key.
    */
-  public RSAPublicKey getX002PublicKey();
+  RSAPublicKey getX002PublicKey();
 
   /**
    * Returns the signature certificate.
    * @return the encryption certificate.
    * @throws EbicsException
    */
-  public byte[] getA005Certificate() throws EbicsException;
+  byte[] getA005Certificate() throws EbicsException;
 
   /**
    * Returns the authentication certificate.
    * @return the encryption certificate.
    * @throws EbicsException
    */
-  public byte[] getX002Certificate() throws EbicsException;
+  byte[] getX002Certificate() throws EbicsException;
 
   /**
    * Returns the encryption certificate.
    * @return the encryption certificate.
    * @throws EbicsException
    */
-  public byte[] getE002Certificate() throws EbicsException;
+  byte[] getE002Certificate() throws EbicsException;
 
   /**
    * Sets the signature certificate.
    * @param a005certificate the signature certificate.
    */
-  public void setA005Certificate(X509Certificate a005certificate);
+  void setA005Certificate(X509Certificate a005certificate);
 
   /**
    * Sets the authentication certificate.
-   * @param a005certificate the authentication certificate.
+   * @param x002certificate the authentication certificate.
    */
-  public void setX002Certificate(X509Certificate x002certificate);
+  void setX002Certificate(X509Certificate x002certificate);
 
   /**
    * Sets the encryption certificate.
-   * @param a005certificate the encryption certificate.
+   * @param e002certificate the encryption certificate.
    */
-  public void setE002Certificate(X509Certificate e002certificate);
+  void setE002Certificate(X509Certificate e002certificate);
 
   /**
    * Sets the signature private key
    * @param a005Key the signature private key
    */
-  public void setA005PrivateKey(PrivateKey a005Key);
+  void setA005PrivateKey(PrivateKey a005Key);
 
   /**
    * Sets the authentication private key
-   * @param a005Key the authentication private key
+   * @param x002Key the authentication private key
    */
-  public void setX002PrivateKey(PrivateKey x002Key);
+  void setX002PrivateKey(PrivateKey x002Key);
 
   /**
    * Sets the encryption private key
-   * @param a005Key the encryption private key
+   * @param e002Key the encryption private key
    */
-  public void setE002PrivateKey(PrivateKey e002Key);
+  void setE002PrivateKey(PrivateKey e002Key);
 
   /**
    * Returns the type to security medium used to store the A005 key.
    * @return the type to security medium used to store the A005 key.
    */
-  public String getSecurityMedium();
+  String getSecurityMedium();
 
   /**
    * Returns the customer in whose name we operate.
    * @return the customer in whose name we operate.
    */
-  public EbicsPartner getPartner();
+  EbicsPartner getPartner();
 
   /**
    * Returns the (bank provided) user id.
    * @return the (bank provided) user id.
    */
-  public String getUserId();
+  String getUserId();
 
   /**
    * Returns the user name.
    * @return the user name.
    */
-  public String getName();
+  String getName();
 
   /**
    * Returns the distinguished name
    * @return the distinguished name
    */
-  public String getDN();
+  String getDN();
 
   /**
    * Returns the password callback handler for the current user.
    * @return the password callback handler.
    */
-  public PasswordCallback getPasswordCallback();
+  PasswordCallback getPasswordCallback();
 
   /**
    * Signs the given digest with the private X002 key.
@@ -153,7 +151,7 @@ public interface EbicsUser {
    * @return the signature.
    * @throws GeneralSecurityException
    */
-  public byte[] authenticate(byte[] digest) throws GeneralSecurityException;
+  byte[] authenticate(byte[] digest) throws GeneralSecurityException;
 
   /**
    * Signs the given digest with the private A005 key.
@@ -162,7 +160,7 @@ public interface EbicsUser {
    * @throws IOException
    * @throws GeneralSecurityException 
    */
-  public byte[] sign(byte[] digest) throws IOException, GeneralSecurityException;
+  byte[] sign(byte[] digest) throws IOException, GeneralSecurityException;
 
   /**
    * Uses the E001 key to decrypt the given secret key.
@@ -173,6 +171,6 @@ public interface EbicsUser {
    * @throws IOException
    * @throws EbicsException 
    */
-  public byte[] decrypt(byte[] encryptedKey, byte[] transactionKey)
+  byte[] decrypt(byte[] encryptedKey, byte[] transactionKey)
     throws GeneralSecurityException, IOException, EbicsException;
 }
