@@ -14,11 +14,11 @@
  * License along with this library; if not, write to the Free Software
  * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA 02111-1307 USA
  *
- * $Id$
  */
 
 package org.kopi.ebics.interfaces;
 
+import java.io.File;
 import java.io.ObjectInputStream;
 
 import org.kopi.ebics.exception.EbicsException;
@@ -29,7 +29,6 @@ import org.kopi.ebics.exception.EbicsException;
  * The manager should ensure serialization and deserialization
  * operations
  *
- * @author hachani
  *
  */
 public interface SerializationManager {
@@ -39,7 +38,7 @@ public interface SerializationManager {
    * @param object the <code>Savable</code> object$
    * @throws EbicsException serialization fails
    */
-  public void serialize(Savable object) throws EbicsException;
+  void serialize(Savable object) throws EbicsException;
 
   /**
    * Deserializes the given object input stream.
@@ -47,11 +46,11 @@ public interface SerializationManager {
    * @return the corresponding object input stream
    * @throws EbicsException deserialization fails
    */
-  public ObjectInputStream deserialize(String name) throws EbicsException;
+  ObjectInputStream deserialize(String name) throws EbicsException;
 
   /**
    * Sets the serialization directory
    * @param serializationDir the serialization directory
    */
-  public void setSerializationDirectory(String serializationDir);
+  void setSerializationDirectory(File serializationDir);
 }
