@@ -14,7 +14,6 @@
  * License along with this library; if not, write to the Free Software
  * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA 02111-1307 USA
  *
- * $Id$
  */
 
 package org.kopi.ebics.io;
@@ -22,6 +21,7 @@ package org.kopi.ebics.io;
 import java.io.ByteArrayInputStream;
 import java.io.IOException;
 import java.io.InputStream;
+import java.nio.charset.StandardCharsets;
 
 import org.kopi.ebics.interfaces.ContentFactory;
 
@@ -31,7 +31,6 @@ import org.kopi.ebics.interfaces.ContentFactory;
  * as a <code>ByteArrayInputStream</code>. This object is
  * serializable in a way to recover interrupted file transfers.
  *
- * @author hachani
  *
  */
 public class ByteArrayContentFactory implements ContentFactory {
@@ -50,10 +49,16 @@ public class ByteArrayContentFactory implements ContentFactory {
     return new ByteArrayInputStream(content);
   }
 
+
+  @Override
+  public String toString() {
+    return new String(content, StandardCharsets.UTF_8);
+  }
+
   // --------------------------------------------------------------------
   // DATA MEMBERS
   // --------------------------------------------------------------------
 
-  private byte[]		content;
+  private final byte[]		content;
   private static final long 	serialVersionUID = 4708928447888655116L;
 }
