@@ -14,7 +14,6 @@
  * License along with this library; if not, write to the Free Software
  * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA 02111-1307 USA
  *
- * $Id$
  */
 
 package org.kopi.ebics.io;
@@ -26,7 +25,6 @@ import java.util.Map;
 /**
  * A simple mean to cache created files.
  *
- * @author hachani
  *
  */
 public class FileCache {
@@ -37,7 +35,7 @@ public class FileCache {
    */
   public FileCache(boolean isTraceEnabled) {
     this.isTraceEnabled = isTraceEnabled;
-    cache = new Hashtable<String, File>();
+    cache = new Hashtable<>();
   }
 
   /**
@@ -95,6 +93,6 @@ public class FileCache {
   // DATA MEMBERS
   // --------------------------------------------------------------------
 
-  private Map<String, File>			cache;
+  private final Map<String, File>			cache;
   private boolean 				isTraceEnabled;
 }
