@@ -14,7 +14,6 @@
  * License along with this library; if not, write to the Free Software
  * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA 02111-1307 USA
  *
- * $Id$
  */
 
 package org.kopi.ebics.io;
@@ -31,7 +30,6 @@ import org.kopi.ebics.interfaces.ContentFactory;
  * This object is serializable in a way to recover interrupted
  * file transfers.
  *
- * @author hachani
  *
  */
 public class InputStreamContentFactory implements ContentFactory {
@@ -53,6 +51,6 @@ public class InputStreamContentFactory implements ContentFactory {
   // DATA MEMBERS
   // --------------------------------------------------------------------
 
-  private InputStream			input;
+  private final InputStream			input;
   private static final long 		serialVersionUID = 2357104115203917168L;
 }
