@@ -14,7 +14,6 @@
  * License along with this library; if not, write to the Free Software
  * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA 02111-1307 USA
  *
- * $Id$
  */
 
 package org.kopi.ebics.io;
@@ -113,7 +112,7 @@ public class Splitter {
    * a given segment number.
    * 
    * @param segmentNumber the segment number
-   * @return
+   * @return content of a data segment
    */
   public ContentFactory getContent(int segmentNumber) {
     byte[]		segment;
