@@ -14,7 +14,6 @@
  * License along with this library; if not, write to the Free Software
  * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA 02111-1307 USA
  *
- * $Id$
  */
 
 package org.kopi.ebics.security;
@@ -25,7 +24,6 @@ import org.kopi.ebics.interfaces.PasswordCallback;
  * A simple user password handler that creates a password
  * from the user id and a simple suffix.
  *
- * @author hachani
  *
  */
 public class UserPasswordHandler implements PasswordCallback {
@@ -53,6 +51,6 @@ public class UserPasswordHandler implements PasswordCallback {
   // DATA MEMBERS
   // --------------------------------------------------------------------
 
-  private String		userId;
-  private String		suffix;
+  private final String		userId;
+  private final String		suffix;
 }
