@@ -14,7 +14,6 @@
  * License along with this library; if not, write to the Free Software
  * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA 02111-1307 USA
  *
- * $Id$
  */
 
 package org.kopi.ebics.xml;
@@ -23,6 +22,7 @@ import java.io.IOException;
 
 import org.apache.xmlbeans.XmlException;
 import org.apache.xmlbeans.XmlObject;
+import org.apache.xmlbeans.impl.schema.DocumentFactory;
 import org.kopi.ebics.exception.EbicsException;
 import org.kopi.ebics.exception.ReturnCode;
 import org.kopi.ebics.interfaces.ContentFactory;
@@ -32,7 +32,6 @@ import org.kopi.ebics.interfaces.ContentFactory;
  * The <code>DefaultResponseElement</code> is the common element for
  * all ebics server responses.
  *
- * @author Hachani
  *
  */
 public abstract class DefaultResponseElement extends DefaultEbicsRootElement {
@@ -42,26 +41,28 @@ public abstract class DefaultResponseElement extends DefaultEbicsRootElement {
    * @param factory the content factory containing the response.
    * @param name the element name
    */
-  public DefaultResponseElement(ContentFactory factory, String name) {
+  protected DefaultResponseElement(ContentFactory factory, String name) {
     this.factory = factory;
     this.name = name;
   }
 
   /**
    * Parses the content of a <code>ContentFactory</code>
+   *
    * @param factory the content factory
+   * @return the parsed document
    * @throws EbicsException parse error
    */
-  protected void parse(ContentFactory factory) throws EbicsException {
-    try {
-      document = XmlObject.Factory.parse(factory.getContent());
-    } catch (XmlException e) {
-      throw new EbicsException(e.getMessage());
-    } catch (IOException e) {
-      throw new EbicsException(e.getMessage());
-    }
+  protected <T extends XmlObject> T parse(ContentFactory factory,
+      DocumentFactory<T> documentFactory) throws EbicsException {
+      try {
+          var doc = documentFactory.parse(factory.getContent());
+          this.document = doc;
+          return doc;
+      } catch (XmlException | IOException e) {
+          throw new EbicsException(e.getMessage());
+      }
   }
-
   /**
    * Reports the return code to the user.
    * @throws EbicsException request fails.
@@ -85,7 +86,7 @@ public abstract class DefaultResponseElement extends DefaultEbicsRootElement {
   // DATA MEMBERS
   // --------------------------------------------------------------------
 
-  private String 			name;
+  private final String 			name;
   protected ContentFactory		factory;
   protected ReturnCode			returnCode;
   private static final long 		serialVersionUID = 4014595046719645090L;
