@@ -14,33 +14,26 @@
  * License along with this library; if not, write to the Free Software
  * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA 02111-1307 USA
  *
- * $Id$
  */
 
 package org.kopi.ebics.xml;
 
 import java.util.Calendar;
-import java.util.Date;
 
 import org.kopi.ebics.exception.EbicsException;
-import org.kopi.ebics.schema.h003.EbicsRequestDocument.EbicsRequest;
-import org.kopi.ebics.schema.h003.EbicsRequestDocument.EbicsRequest.Body;
-import org.kopi.ebics.schema.h003.EbicsRequestDocument.EbicsRequest.Header;
-import org.kopi.ebics.schema.h003.FDLOrderParamsType;
-import org.kopi.ebics.schema.h003.FDLOrderParamsType.DateRange;
-import org.kopi.ebics.schema.h003.FileFormatType;
-import org.kopi.ebics.schema.h003.MutableHeaderType;
-import org.kopi.ebics.schema.h003.OrderAttributeType;
-import org.kopi.ebics.schema.h003.ParameterDocument.Parameter;
-import org.kopi.ebics.schema.h003.ParameterDocument.Parameter.Value;
-import org.kopi.ebics.schema.h003.StandardOrderParamsType;
-import org.kopi.ebics.schema.h003.StaticHeaderOrderDetailsType;
-import org.kopi.ebics.schema.h003.StaticHeaderOrderDetailsType.OrderType;
-import org.kopi.ebics.schema.h003.StaticHeaderType;
-import org.kopi.ebics.schema.h003.StaticHeaderType.BankPubKeyDigests;
-import org.kopi.ebics.schema.h003.StaticHeaderType.BankPubKeyDigests.Authentication;
-import org.kopi.ebics.schema.h003.StaticHeaderType.BankPubKeyDigests.Encryption;
-import org.kopi.ebics.schema.h003.StaticHeaderType.Product;
+import org.kopi.ebics.interfaces.EbicsOrderType;
+import org.kopi.ebics.schema.h005.EbicsRequestDocument.EbicsRequest;
+import org.kopi.ebics.schema.h005.EbicsRequestDocument.EbicsRequest.Body;
+import org.kopi.ebics.schema.h005.EbicsRequestDocument.EbicsRequest.Header;
+import org.kopi.ebics.schema.h005.MutableHeaderType;
+import org.kopi.ebics.schema.h005.StandardOrderParamsDocument;
+import org.kopi.ebics.schema.h005.StandardOrderParamsType;
+import org.kopi.ebics.schema.h005.StaticHeaderOrderDetailsType;
+import org.kopi.ebics.schema.h005.StaticHeaderType;
+import org.kopi.ebics.schema.h005.StaticHeaderType.BankPubKeyDigests;
+import org.kopi.ebics.schema.h005.StaticHeaderType.BankPubKeyDigests.Authentication;
+import org.kopi.ebics.schema.h005.StaticHeaderType.BankPubKeyDigests.Encryption;
+import org.kopi.ebics.schema.h005.StaticHeaderType.Product;
 import org.kopi.ebics.session.EbicsSession;
 
 
@@ -48,7 +41,6 @@ import org.kopi.ebics.session.EbicsSession;
  * The <code>DInitializationRequestElement</code> is the common initialization
  * for all ebics downloads.
  *
- * @author Hachani
  *
  */
 public class DownloadInitializationRequestElement extends InitializationRequestElement {
@@ -57,19 +49,10 @@ public class DownloadInitializationRequestElement extends InitializationRequestE
    * Constructs a new <code>DInitializationRequestElement</code> for downloads initializations.
    * @param session the current ebics session
    * @param type the download order type (FDL, HTD, HPD)
-   * @param startRange the start range download
-   * @param endRange the end range download
-   * @throws EbicsException
    */
   public DownloadInitializationRequestElement(EbicsSession session,
-                                       org.kopi.ebics.session.OrderType type,
-                                       Date startRange,
-                                       Date endRange)
-    throws EbicsException
-  {
+                                       EbicsOrderType type) {
     super(session, type, generateName(type));
-    this.startRange = startRange;
-    this.endRange = endRange;
   }
 
   @Override
@@ -83,7 +66,6 @@ public class DownloadInitializationRequestElement extends InitializationRequestE
     BankPubKeyDigests 			bankPubKeyDigests;
     Authentication 			authentication;
     Encryption 				encryption;
-    OrderType 				orderType;
     StaticHeaderOrderDetailsType 	orderDetails;
 
     mutable = EbicsXmlFactory.createMutableHeaderType("Initialisation", null);
@@ -95,44 +77,18 @@ public class DownloadInitializationRequestElement extends InitializationRequestE
 	                                          "http://www.w3.org/2001/04/xmlenc#sha256",
 	                                          decodeHex(session.getUser().getPartner().getBank().getE002Digest()));
     bankPubKeyDigests = EbicsXmlFactory.createBankPubKeyDigests(authentication, encryption);
-    orderType = EbicsXmlFactory.createOrderType(type.toString());
-    if (type.equals(org.kopi.ebics.session.OrderType.FDL)) {
-      FDLOrderParamsType		fDLOrderParamsType;
-      FileFormatType 			fileFormat;
 
-      fileFormat = EbicsXmlFactory.createFileFormatType(session.getConfiguration().getLocale().getCountry().toUpperCase(),
-	                                                session.getSessionParam("FORMAT"));
-      fDLOrderParamsType = EbicsXmlFactory.createFDLOrderParamsType(fileFormat);
+      StandardOrderParamsType standardOrderParamsType = EbicsXmlFactory.createStandardOrderParamsType();
 
-      if (startRange != null && endRange != null) {
-	DateRange		range;
+      var type = StaticHeaderOrderDetailsType.AdminOrderType.Factory.newInstance();
+      type.setStringValue(this.getType());
 
-	range = EbicsXmlFactory.createDateRange(startRange, endRange);
-	fDLOrderParamsType.setDateRange(range);
-      }
-
-      if (Boolean.getBoolean(session.getSessionParam("TEST"))) {
-	Parameter 		parameter;
-	Value			value;
-
-	value = EbicsXmlFactory.createValue("String", "TRUE");
-	parameter = EbicsXmlFactory.createParameter("TEST", value);
-	fDLOrderParamsType.setParameterArray(new Parameter[] {parameter});
-      }
-      orderDetails = EbicsXmlFactory.createStaticHeaderOrderDetailsType(session.getUser().getPartner().nextOrderId(),
-                                                                        OrderAttributeType.DZHNN,
-                                                                        orderType,
-                                                                        fDLOrderParamsType);
-    } else {
-      StandardOrderParamsType		standardOrderParamsType;
-
-      standardOrderParamsType = EbicsXmlFactory.createStandardOrderParamsType();
       //FIXME Some banks cannot handle OrderID element in download process. Add parameter in configuration!!!
       orderDetails = EbicsXmlFactory.createStaticHeaderOrderDetailsType(null,//session.getUser().getPartner().nextOrderId(),
-	                                                                OrderAttributeType.DZHNN,
-	                                                                orderType,
-	                                                                standardOrderParamsType);
-    }
+            type,
+	                                                                standardOrderParamsType,
+          StandardOrderParamsDocument.type);
+
     xstatic = EbicsXmlFactory.createStaticHeaderType(session.getBankID(),
                                                      nonce,
                                                      session.getUser().getPartner().getPartnerId(),
@@ -151,11 +107,5 @@ public class DownloadInitializationRequestElement extends InitializationRequestE
     document = EbicsXmlFactory.createEbicsRequestDocument(request);
   }
 
-  // --------------------------------------------------------------------
-  // DATA MEMBERS
-  // --------------------------------------------------------------------
-
-  private Date					startRange;
-  private Date					endRange;
   private static final long 			serialVersionUID = 3776072549761880272L;
 }
