@@ -14,7 +14,6 @@
  * License along with this library; if not, write to the Free Software
  * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA 02111-1307 USA
  *
- * $Id$
  */
 
 package org.kopi.ebics.xml;
@@ -23,13 +22,12 @@ import org.kopi.ebics.exception.EbicsException;
 import org.kopi.ebics.exception.NoDownloadDataAvailableException;
 import org.kopi.ebics.exception.ReturnCode;
 import org.kopi.ebics.interfaces.ContentFactory;
-import org.kopi.ebics.session.OrderType;
+import org.kopi.ebics.interfaces.EbicsOrderType;
 
 /**
  * The <code>DInitializationResponseElement</code> is the response element
  * for ebics downloads initializations.
  *
- * @author Hachani
  *
  */
 public class DownloadInitializationResponseElement extends InitializationResponseElement {
@@ -41,7 +39,7 @@ public class DownloadInitializationResponseElement extends InitializationRespons
    * @param name the element name
    */
   public DownloadInitializationResponseElement(ContentFactory factory,
-                                        OrderType orderType,
+                                        EbicsOrderType orderType,
                                         String name)
   {
     super(factory, orderType, name);
