@@ -14,20 +14,18 @@
  * License along with this library; if not, write to the Free Software
  * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA 02111-1307 USA
  *
- * $Id$
  */
 
 package org.kopi.ebics.xml;
 
 import org.kopi.ebics.exception.EbicsException;
 import org.kopi.ebics.interfaces.ContentFactory;
-import org.kopi.ebics.session.OrderType;
+import org.kopi.ebics.interfaces.EbicsOrderType;
 
 /**
  * The <code>DTransferResponseElement</code> is the response element
  * for all ebics downloads transfers.
  *
- * @author Hachani
  *
  */
 public class DownloadTransferResponseElement extends TransferResponseElement {
@@ -39,7 +37,7 @@ public class DownloadTransferResponseElement extends TransferResponseElement {
    * @param name the element name.
    */
   public DownloadTransferResponseElement(ContentFactory factory,
-                                  OrderType orderType,
+                                  EbicsOrderType orderType,
                                   String name)
   {
     super(factory, name);
