@@ -14,7 +14,6 @@
  * License along with this library; if not, write to the Free Software
  * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA 02111-1307 USA
  *
- * $Id$
  */
 
 package org.kopi.ebics.xml;
@@ -29,7 +28,6 @@ import org.kopi.ebics.utils.Utils;
  * to send the authentication and encryption keys to the ebics
  * bank server
  *
- * @author hachani
  *
  */
 public class HIARequestElement extends DefaultEbicsRootElement {
@@ -51,23 +49,21 @@ public class HIARequestElement extends DefaultEbicsRootElement {
 
   @Override
   public void build() throws EbicsException {
-    HIARequestOrderDataElement		requestOrderData;
 
-    requestOrderData = new HIARequestOrderDataElement(session);
-    requestOrderData.build();
-    unsecuredRequest = new UnsecuredRequestElement(session,
-	                                           OrderType.HIA,
-	                                           orderId == null ? session.getUser().getPartner().nextOrderId() : orderId,
-	                                           Utils.zip(requestOrderData.prettyPrint()));
-    unsecuredRequest.build();
+      var requestOrderData = new HIARequestOrderDataElement(session);
+      requestOrderData.build();
+      unsecuredRequest = new UnsecuredRequestElement(session, OrderType.HIA,
+          orderId == null ? session.getUser().getPartner().nextOrderId() : orderId,
+          Utils.zip(requestOrderData.prettyPrint()));
+      unsecuredRequest.setSaveSuggestedPrefixes("urn:org:ebics:H005", "");
+      unsecuredRequest.build();
   }
 
-  @Override
-  public byte[] toByteArray() {
-    setSaveSuggestedPrefixes("http://www.ebics.org/H003", "");
+    @Override
+    public byte[] toByteArray() {
 
-    return unsecuredRequest.toByteArray();
-  }
+        return unsecuredRequest.toByteArray();
+    }
 
   @Override
   public void validate() throws EbicsException {
@@ -78,7 +74,7 @@ public class HIARequestElement extends DefaultEbicsRootElement {
   // DATA MEMBERS
   // --------------------------------------------------------------------
 
-  private String			orderId;
+  private final String			orderId;
   private UnsecuredRequestElement	unsecuredRequest;
   private static final long 		serialVersionUID = 1130436605993828777L;
 }
