@@ -14,7 +14,6 @@
  * License along with this library; if not, write to the Free Software
  * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA 02111-1307 USA
  *
- * $Id$
  */
 
 package org.kopi.ebics.xml;
@@ -22,15 +21,14 @@ package org.kopi.ebics.xml;
 import org.kopi.ebics.exception.EbicsException;
 import org.kopi.ebics.exception.ReturnCode;
 import org.kopi.ebics.interfaces.ContentFactory;
-import org.kopi.ebics.schema.h003.EbicsResponseDocument;
-import org.kopi.ebics.schema.h003.EbicsResponseDocument.EbicsResponse;
-import org.kopi.ebics.session.OrderType;
+import org.kopi.ebics.interfaces.EbicsOrderType;
+import org.kopi.ebics.schema.h005.EbicsResponseDocument;
+import org.kopi.ebics.schema.h005.EbicsResponseDocument.EbicsResponse;
 
 /**
  * The <code>InitializationResponseElement</code> is the common
  * element for transfer initialization responses.
  *
- * @author Hachani
  *
  */
 public class InitializationResponseElement extends DefaultResponseElement {
@@ -42,24 +40,24 @@ public class InitializationResponseElement extends DefaultResponseElement {
    * @param name the element name
    */
   public InitializationResponseElement(ContentFactory factory,
-                                       OrderType orderType,
+                                       EbicsOrderType orderType,
                                        String name)
   {
     super(factory, name);
     this.orderType = orderType;
   }
 
-  @Override
-  public void build() throws EbicsException {
-    parse(factory);
-    response = ((EbicsResponseDocument)document).getEbicsResponse();
-    String code = response.getHeader().getMutable().getReturnCode();
-    String text = response.getHeader().getMutable().getReportText();
-    returnCode = ReturnCode.toReturnCode(code, text);
-    checkReturnCode(returnCode);
-    processBodyReturnCode();
-    transactionId = response.getHeader().getStatic().getTransactionID();
-  }
+    @Override
+    public void build() throws EbicsException {
+        var doc = parse(factory, EbicsResponseDocument.Factory);
+        response = doc.getEbicsResponse();
+        String code = response.getHeader().getMutable().getReturnCode();
+        String text = response.getHeader().getMutable().getReportText();
+        returnCode = ReturnCode.toReturnCode(code, text);
+        checkReturnCode(returnCode);
+        processBodyReturnCode();
+        transactionId = response.getHeader().getStatic().getTransactionID();
+    }
 
   protected void processBodyReturnCode() throws EbicsException {
       String bodyRetCode = response.getBody().getReturnCode().getStringValue();
@@ -81,7 +79,7 @@ public class InitializationResponseElement extends DefaultResponseElement {
    * @return the order type.
    */
   public String getOrderType() {
-    return orderType.toString();
+    return orderType.getCode();
   }
 
   // --------------------------------------------------------------------
@@ -89,7 +87,7 @@ public class InitializationResponseElement extends DefaultResponseElement {
   // --------------------------------------------------------------------
 
   protected EbicsResponse			response;
-  private OrderType				orderType;
+  private final EbicsOrderType orderType;
   private byte[]				transactionId;
   private static final long 			serialVersionUID = 7684048385353175772L;
 }
