@@ -14,7 +14,6 @@
  * License along with this library; if not, write to the Free Software
  * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA 02111-1307 USA
  *
- * $Id$
  */
 
 package org.kopi.ebics.xml;
@@ -22,14 +21,12 @@ package org.kopi.ebics.xml;
 import org.kopi.ebics.exception.EbicsException;
 import org.kopi.ebics.exception.ReturnCode;
 import org.kopi.ebics.interfaces.ContentFactory;
-import org.kopi.ebics.schema.h003.EbicsResponseDocument;
-import org.kopi.ebics.schema.h003.EbicsResponseDocument.EbicsResponse;
+import org.kopi.ebics.schema.h005.EbicsResponseDocument;
 
 /**
  * The <code>ReceiptResponseElement</code> is the response element
  * for ebics receipt request.
  *
- * @author Hachani
  *
  */
 public class ReceiptResponseElement extends DefaultResponseElement {
@@ -45,16 +42,12 @@ public class ReceiptResponseElement extends DefaultResponseElement {
 
   @Override
   public void build() throws EbicsException {
-    String			code;
-    String			text;
-    EbicsResponse		response;
-
-    parse(factory);
-    response = ((EbicsResponseDocument)document).getEbicsResponse();
-    code = response.getHeader().getMutable().getReturnCode();
-    text = response.getHeader().getMutable().getReportText();
-    returnCode = ReturnCode.toReturnCode(code, text);
-    report();
+      var doc = parse(factory, EbicsResponseDocument.Factory);
+      var response = doc.getEbicsResponse();
+      String code = response.getHeader().getMutable().getReturnCode();
+      String text = response.getHeader().getMutable().getReportText();
+      returnCode = ReturnCode.toReturnCode(code, text);
+      report();
   }
 
   @Override
