@@ -14,20 +14,21 @@
  * License along with this library; if not, write to the Free Software
  * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA 02111-1307 USA
  *
- * $Id$
  */
 
 package org.kopi.ebics.xml;
 
 import java.io.ByteArrayInputStream;
+import java.io.ByteArrayOutputStream;
 
 import javax.xml.parsers.DocumentBuilder;
 import javax.xml.parsers.DocumentBuilderFactory;
+import javax.xml.xpath.XPathConstants;
+import javax.xml.xpath.XPathFactory;
 
 import org.apache.xml.security.c14n.Canonicalizer;
 import org.apache.xml.security.transforms.TransformationException;
 import org.apache.xml.security.utils.IgnoreAllErrorHandler;
-import org.apache.xpath.XPathAPI;
 import org.kopi.ebics.exception.EbicsException;
 import org.kopi.ebics.interfaces.EbicsUser;
 import org.kopi.ebics.schema.xmldsig.CanonicalizationMethodType;
@@ -101,7 +102,6 @@ public class SignedInfo extends DefaultEbicsRootElement {
   /**
    * Returns the signed info element as an <code>XmlObject</code>
    * @return he signed info element
-   * @throws EbicsException user Signature and Canonicalization errors
    */
   public SignatureType getSignatureType() {
     return ((SignatureType)document);
@@ -133,7 +133,6 @@ public class SignedInfo extends DefaultEbicsRootElement {
       DocumentBuilderFactory 		factory;
       DocumentBuilder			builder;
       Document				document;
-      Node 				node;
       Canonicalizer 			canonicalizer;
 
       factory = DocumentBuilderFactory.newInstance();
@@ -142,9 +141,12 @@ public class SignedInfo extends DefaultEbicsRootElement {
       builder = factory.newDocumentBuilder();
       builder.setErrorHandler(new IgnoreAllErrorHandler());
       document = builder.parse(new ByteArrayInputStream(toSign));
-      node = XPathAPI.selectSingleNode(document, "//ds:SignedInfo");
+      Node node = (Node) XPathFactory.newInstance().newXPath()
+           .evaluate("//*[name()='ds:SignedInfo']", document, XPathConstants.NODE);
       canonicalizer = Canonicalizer.getInstance(Canonicalizer.ALGO_ID_C14N_OMIT_COMMENTS);
-      return user.authenticate(canonicalizer.canonicalizeSubtree(node));
+      var bos = new ByteArrayOutputStream();
+      canonicalizer.canonicalizeSubtree(node, bos);
+      return user.authenticate(bos.toByteArray());
     } catch(Exception e) {
       throw new EbicsException(e.getMessage());
     }
@@ -167,7 +169,7 @@ public class SignedInfo extends DefaultEbicsRootElement {
   // DATA MEMBERS
   // --------------------------------------------------------------------
 
-  private byte[]			digest;
-  private EbicsUser 			user;
+  private final byte[]			digest;
+  private final EbicsUser 			user;
   private static final long 		serialVersionUID = 4194924578678778580L;
 }
