@@ -14,7 +14,6 @@
  * License along with this library; if not, write to the Free Software
  * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA 02111-1307 USA
  *
- * $Id$
  */
 
 package org.kopi.ebics.xml;
@@ -22,14 +21,13 @@ package org.kopi.ebics.xml;
 import org.kopi.ebics.exception.EbicsException;
 import org.kopi.ebics.exception.ReturnCode;
 import org.kopi.ebics.interfaces.ContentFactory;
-import org.kopi.ebics.schema.h003.EbicsResponseDocument;
-import org.kopi.ebics.schema.h003.EbicsResponseDocument.EbicsResponse;
+import org.kopi.ebics.schema.h005.EbicsResponseDocument;
+import org.kopi.ebics.schema.h005.EbicsResponseDocument.EbicsResponse;
 
 /**
  * The <code>TransferResponseElement</code> is the common element
  * response for all ebics transfers.
  *
- * @author Hachani
  *
  */
 public class TransferResponseElement extends DefaultResponseElement {
@@ -37,7 +35,6 @@ public class TransferResponseElement extends DefaultResponseElement {
   /**
    * Constructs a new <code>TransferResponseElement</code> element.
    * @param factory the content factory
-   * @param orderType the order type
    * @param name the element name;
    */
   public TransferResponseElement(ContentFactory factory, String name) {
@@ -46,13 +43,10 @@ public class TransferResponseElement extends DefaultResponseElement {
 
   @Override
   public void build() throws EbicsException {
-    String			code;
-    String			text;
-
-    parse(factory);
-    response = ((EbicsResponseDocument)document).getEbicsResponse();
-    code = response.getHeader().getMutable().getReturnCode();
-    text = response.getHeader().getMutable().getReportText();
+    var doc = parse(factory, EbicsResponseDocument.Factory);
+    response = doc.getEbicsResponse();
+    String code = response.getHeader().getMutable().getReturnCode();
+    String text = response.getHeader().getMutable().getReportText();
     returnCode = ReturnCode.toReturnCode(code, text);
     report();
   }
